pyproject.toml
[build-system]
requires = ["setuptools>=67.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]

[project]
name = "oggmap"
version = "0.0.2"
description = "extract orthologous maps (short: orthomap) from OrthoFinder output for query species"
authors = [
    {name = "Kristian K Ullrich", email="ullrich@evolbio.mpg.de"},
    {name = "Nikoleta E Glynatsi", email="glynatsi@evolbio.mpg.de"},
]
maintainers = [
    {name = "Kristian K Ullrich", email="ullrich@evolbio.mpg.de"},
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["orthomap", "gene age", "orthologs", "phylostratigraphy", "single-cell"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
]
dependencies = [
    "setuptools>=67.6.0",
    "numpy<2.2,>=1.24",
    "numba>=0.61.0",
    "scipy>=1.15.1",
    "pandas>=2.2.3",
    "anndata>=0.11.3",
    "scanpy>=1.11.0",
    "seaborn>=0.13.2",
    "statannot>=0.2.3",
    "harmonypy>=0.0.10",
    "leidenalg>=0.10.2",
    "umap-learn>=0.5.7",
    "alive-progress>=3.2.0",
    "biopython>=1.85",
    "taxadb2>=0.12.3",
    "wget>=3.2",
    "pytest",
    "pip",
]

[project.urls]
"Homepage" = "https://github.com/kullrich/oggmap"
"Bug Tracker" = "https://github.com/kullrich/oggmap/issues"
documentation = "https://oggmap.readthedocs.io/en/latest/"
repository = "https://github.com/kullrich/oggmap"

[project.scripts]
oggmap = "oggmap.__main__:main"
cds2aa = "oggmap.cds2aa:main"
eggnog2orthomap = "oggmap.eggnog2orthomap:main"
plaza2orthomap = "oggmap.plaza2orthomap:main"
orthomcl2orthomap = "oggmap.orthomcl2orthomap:main"
gtf2t2g = "oggmap.gtf2t2g:main"
ncbitax = "oggmap.ncbitax:main"
of2orthomap = "oggmap.of2orthomap:main"
qlin = "oggmap.qlin:main"
